refactor: change user entity and user mapper class
